+++ sherlodoc/index/type_cache.ml
@@ -16,7 +16,8 @@ let rec of_odoc ~cache otyp =
   match otyp with
   | Odoc_model.Lang.TypeExpr.Var _str -> Any
   | Any -> Any
-  | Arrow (_lbl, left, right) -> cache (Arrow (of_odoc ~cache left, of_odoc ~cache right))
+  | Arrow (_lbl, (left, _), (right, _)) ->
+      cache (Arrow (of_odoc ~cache left, of_odoc ~cache right))
   | Constr (name, args) ->
       cache (Constr (Typename.to_string name, List.map (of_odoc ~cache) args))
   | Tuple li -> cache (Tuple (List.map (fun (_, ty) -> of_odoc ~cache ty) li))