+++ src/xref2/subst.ml
@@ -164,8 +164,8 @@ let rec substitute_vars vars t =
   | Var s -> ( try List.assoc s vars with Not_found -> t)
   | Any -> Any
   | Alias (t, str) -> Alias (substitute_vars vars t, str)
-  | Arrow (lbl, t1, t2) ->
-      Arrow (lbl, substitute_vars vars t1, substitute_vars vars t2)
+  | Arrow (lbl, (t1, m1), (t2, m2)) ->
+      Arrow (lbl, (substitute_vars vars t1, m1), (substitute_vars vars t2, m2))
   | Tuple ts ->
       Tuple (List.map (fun (lbl, ty) -> (lbl, substitute_vars vars ty)) ts)
   | Unboxed_tuple ts ->
@@ -609,7 +609,8 @@ and type_expr s t =
   | Var s -> Var s
   | Any -> Any
   | Alias (t, str) -> Alias (type_expr s t, str)
-  | Arrow (lbl, t1, t2) -> Arrow (lbl, type_expr s t1, type_expr s t2)
+  | Arrow (lbl, (t1, m1), (t2, m2)) ->
+      Arrow (lbl, (type_expr s t1, m1), (type_expr s t2, m2))
   | Tuple ts -> Tuple (List.map (fun (lbl, ty) -> (lbl, type_expr s ty)) ts)
   | Unboxed_tuple ts ->
       Unboxed_tuple (List.map (fun (l, t) -> (l, type_expr s t)) ts)